ctplib/posixq.c used in ctp_proxy/main_ctp.c

+++ v/vme/ctp_proxy/main_ctp.c
@@ -3,6 +3,7 @@
 strcpy(errorReason,"not set"); smi_set_parER();
 22.11.
 smi_setState() introduced
+4.8.2014: executectp() added
 */
 #include <stdlib.h>
 #include <unistd.h>
@@ -23,6 +24,8 @@ smi_setState() introduced
 #include "Tpartition.h"
 #include "ctp_proxy.h"
 
+#define PQWAY
+//#undef PQWAY
 void ds_register();
 
 char errorReason[ERRMSGL];
@@ -32,8 +35,25 @@ char obj[128];
 int quit=0;   // 1: do not start new runs >9 stop immediately   
 
 void UPPER(char *strin);
+#ifdef PQWAY
+// see ctplib/posixq.c
+#include <mqueue.h>
+mqd_t mq_sendmsg;
+mqd_t mq_rec;
+mqd_t pq_open();
+mqd_t pq_connect();
+void pq_close(mqd_t mq, int unlink);
+int pq_send(mqd_t mq, char *msg);
+void pq_receive(mqd_t mq, char *msg);
+
+#endif
 char pname[64]="UNKNOWN"; 
 char ORBIT_NUMBER[12]=""; 
+char mask[64];
+char ACT_CONFIG[8]="YES";
+int run_number=0;
+int detectors=0;   // ECS bits set for detectors in DETECTORS SMI par
+
 /*---------------------------------------------*/ void gotsignal(int signum) {
 char msg[100];
 // SIGUSR1:  // kill -s USR1 pid
@@ -41,9 +61,11 @@ signal(signum, gotsignal); siginterrupt(signum, 0);
 sprintf(msg, "got signal:%d", signum); prtLog(msg);
 if((signum==SIGUSR1) || (signum==SIGQUIT) ) {
   int np;
-  if((np= getNAllPartitions())) {
+  np= getNAllPartitions();
+  if(np!=0) {
     quit=1; // wait till all partitions stopped
-    sprintf(msg, "Waiting for the stop of all partitions before exit"); prtLog(msg);
+    sprintf(msg, "Waiting for the stop of all partitions before exit");
+    prtLog(msg);
   } else {
     quit=10;  // stop immediately (no partitions loaded)
   };
@@ -51,6 +73,11 @@ if((signum==SIGUSR1) || (signum==SIGQUIT) ) {
     sprintf(msg, "SIGINT, immediate stop"); prtLog(msg);
   quit=11; // =9. stop immediately
 };
+#ifdef PQWAY
+if(quit>9) {
+  pq_send(mq_sendmsg,"quit");
+};
+#endif
 }
 
 void smi_setState(char *newstate) {
@@ -64,14 +91,66 @@ void smi_set_parEF(char *pname) {
 smi_set_par("EXECUTING_FOR",pname,STRING); smi_setState("EXECUTING");
 }
 
+void executectp(char *cmd) {
+int rc, errorway=0;
+#ifdef PQWAY
+if(strcmp(cmd,"wait")==0) {
+  pq_receive(mq_rec, cmd);
+} else {
+  pq_send(mq_sendmsg, cmd);
+  return;
+};
+#endif
+if(strcmp(cmd,"LOAD")==0) {
+  printf("%s partition: %s mask: %s run:%d\n", cmd,pname, mask,run_number);
+  rc= ctp_LoadPartition(pname,mask,run_number,ACT_CONFIG, errorReason);
+} else if(strcmp(cmd,"INIT")==0) {
+  printf("%s partition: %s mask: %s run:%d\n", cmd,pname, mask,run_number);
+  rc= ctp_InitPartition(pname,mask,run_number,ACT_CONFIG, errorReason);
+} else if(strcmp(cmd,"START_PARTITION")==0) {
+  printf("Starting partition: %s\n", pname);
+  rc= ctp_StartPartition(pname, errorReason);
+} else if(strcmp(cmd,"STOP_PARTITION")==0) {
+  printf("Stopping partition: %s\n", pname);
+  rc= ctp_StopPartition(pname); errorway=1;
+} else if(strcmp(cmd,"PAUSE_PARTITION")==0) {
+  rc= ctp_PausePartition(pname, detectors); errorway= 1;
+} else if(strcmp(cmd,"RESUME_PARTITION")==0) {
+  rc= ctp_ResumePartition(pname, detectors); errorway= 1;
+} else if(strcmp(cmd,"RESUME_PARTITION")==0) {
+  w32 orbitn;
+  rc= ctp_SyncPartition(pname, errorReason, &orbitn);
+  if(rc!=0) {
+    printf("ctp_SyncPartition() rc:%d %s\n", rc,errorReason);
+    smi_set_parER();
+    errorway= 1;
+  } else {
+    sprintf(ORBIT_NUMBER,"%u", orbitn);
+    smi_set_par("ORBIT_NUMBER", ORBIT_NUMBER, STRING);
+  }
+} else {
+  printf("executectp cmd ignored: %s\n", cmd);
+  return;
+};
+if(rc){
+  if(errorway) { // i.e. errorReason set already in smi
+    sleep(1); smi_setState("ERROR");
+  } else {
+    printf("%s rc:%d %s\n", cmd, rc, errorReason);
+    smi_set_parER();
+    sleep(1); smi_setState("LOAD_FAILURE");
+  };
+}else{
+  sleep (1); smi_setState("RUNNING");
+};
+}
 /*--------------------------------------------------- SMI_handle_command()
 */
 void SMI_handle_command() {
 char action[64], param[64], parname[64], msg[256];
-int n_params, ptype, psize, i, run_number=0;
-char mask[64],run_number_str[64];
-char ACT_CONFIG[8]="YES";
-int detectors=0;   // ECS bits set for detectors in DETECTORS SMI par
+char run_number_str[64];
+int n_params, ptype, psize, i;
+
 prtProfTime(NULL);
 smi_get_action(action,&n_params); UPPER(action); strcpy(msg,"");
 if(n_params != 2){
@@ -124,7 +203,7 @@ printf("%s Got action %s with parameters %s\n",obj,action,msg);
    
 if (strcmp(state,"RUNNING") == 0 ) {
   if((strcmp(action,"LOAD_PARTITION") == 0) || (strcmp(action,"INIT_PARTITION") == 0)) {
-    int rc; char INITLOAD[8];
+    char INITLOAD[8];
     if(strcmp(action,"LOAD_PARTITION") == 0) {
       strcpy(INITLOAD,"LOAD");
     } else {
@@ -135,68 +214,23 @@ if (strcmp(state,"RUNNING") == 0 ) {
       infolog_trgboth(LOG_FATAL, msg);
     } else {
       smi_set_parEF(pname);
-      printf("%s partition: %s mask: %s run:%d\n", INITLOAD,pname,
-        mask,run_number);
-      if(strcmp(INITLOAD,"LOAD")==0) {
-        rc= ctp_LoadPartition(pname,mask,run_number,ACT_CONFIG, errorReason);
-      } else {
-        rc= ctp_InitPartition(pname,mask,run_number,ACT_CONFIG, errorReason);
-      };
-      if(rc){
-        printf("ctp_%sPartition() rc:%d %s\n", INITLOAD, rc, errorReason);
-        smi_set_parER();
-        sleep(1); smi_setState("LOAD_FAILURE");
-      }else{
-        sleep (1); smi_setState("RUNNING");
-      }
+      executectp(INITLOAD);
     };
   } else if (strcmp(action,"START_PARTITION") == 0) {
-    int rc;
     smi_set_parEF(pname);
-    printf("Starting partition: %s\n", pname);
-    rc= ctp_StartPartition(pname, errorReason);
-    if(rc){
-      printf("ctp_StartPartition() rc:%d %s\n", rc,errorReason);
-      smi_set_parER();
-      sleep(1); smi_setState("LOAD_FAILURE");
-    }else{
-      sleep (1); smi_setState("RUNNING");
-    }
+    executectp(action);
   } else if (strcmp(action,"STOP_PARTITION") == 0) {
     smi_set_parEF(pname);
-    printf("Send EOD and then unload the partition\n");
-    if(ctp_StopPartition(pname)){
-      sleep(1); smi_setState("ERROR");
-    }else{
-      sleep (1); smi_setState("RUNNING");
-    }
+    executectp(action);
   } else if (strcmp(action,"PAUSE_PARTITION") == 0) {
     smi_set_parEF(pname);
-    if(ctp_PausePartition(pname, detectors)){
-      sleep (1); smi_setState("ERROR");
-    }else{
-      sleep (1); smi_setState("RUNNING");
-    }
-  //} else if (strcmp(action,"SYNCH") == 0) {
+    executectp(action);
   } else if (strncmp(action,"SYNC",4) == 0) {   // correct: strcmp("SYNC")
-    int rc; w32 orbitn;
     smi_set_parEF(pname);
-    if((rc= ctp_SyncPartition(pname, errorReason, &orbitn))!=0){
-      printf("ctp_SyncPartition() rc:%d %s\n", rc,errorReason);
-      smi_set_parER();
-      sleep (1); smi_setState("ERROR");
-    }else{
-      sprintf(ORBIT_NUMBER,"%u", orbitn);
-      smi_set_par("ORBIT_NUMBER", ORBIT_NUMBER, STRING);
-      sleep (1); smi_setState("RUNNING");
-    }
+    executectp(action);
   } else if (strcmp(action,"RESUME_PARTITION") == 0) {
     smi_set_parEF(pname);
-    if(ctp_ResumePartition(pname, detectors)){
-      sleep (1); smi_setState("ERROR");
-    } else {
-      sleep (1); smi_setState("RUNNING");
-    }
+    executectp(action);
   } else {
     char msg[200];
     sprintf(msg, "Illegal action %s in state RUNNING ignored.\n",action);
@@ -229,13 +263,25 @@ return;
 }
 /*----------------------------------------*/ int main(int argc, char **argv) {
 int rc;
+infolog_SetFacility("CTP"); infolog_SetStream("",0);
+#ifdef PQWAY
+mq_rec= pq_open();
+if(mq_rec==(mqd_t)-1) {
+  infolog_trgboth(LOG_FATAL, "posix mq_rec not created");
+  exit(8);
+}
+mq_sendmsg= pq_connect();
+if(mq_sendmsg==(mqd_t)-1) {
+  infolog_trgboth(LOG_FATAL, "posix mq_sendmsg not connected");
+  exit(8);
+}
+#endif
 signal(SIGUSR1, gotsignal); siginterrupt(SIGUSR1, 0);
 signal(SIGQUIT, gotsignal); siginterrupt(SIGQUIT, 0);
 signal(SIGKILL, gotsignal); siginterrupt(SIGKILL, 0); // -9
 signal(SIGTERM, gotsignal); siginterrupt(SIGTERM, 0); // kill pid
 signal(SIGINT, gotsignal); siginterrupt(SIGINT, 0);   // CTRL C   2
 partmode[0]='\0';
-infolog_SetFacility("CTP"); infolog_SetStream("",0);
 printf("cshmInit i.e. initBakery(swtriggers/ccread if shm allocated)...\n");
 /*printf("initBakery(swtriggers,4): 0:SOD/EOD 1:gcalib 2:ctp.exe 3:dims\n");
 printf("initBakery(ccread,5): 0:proxy 1:dims 2:ctp+busytool 3:smaq 4:inputs\n");
@@ -263,10 +309,18 @@ strcpy(ORBIT_NUMBER,""); smi_set_par("ORBIT_NUMBER",ORBIT_NUMBER,STRING);
 strcpy(pname,"UNKNOWN"); smi_set_par("EXECUTING_FOR",pname,STRING);
 smi_setState("RUNNING");
 while(1) {
+#ifdef PQWAY
+  executectp("wait");
+#else
   usleep(1000000);
+#endif
   if(quit>9) break;
 };
 rc= ctp_Endproxy();
+#ifdef PQWAY
+pq_close(mq_sendmsg, 0);
+pq_close(mq_rec, 1);
+#endif
 printf("Calling cshmDetach()...\n"); cshmDetach();
 return (0);
 }
